Fixes the file names and class names of the classes due to their roles being modified.
