+++ spec/services/drug_count_importer_spec.rb
@@ -1,26 +1,24 @@
 require File.expand_path '../../spec_helper.rb', __FILE__
-require './src/services/drug_counts_importer.rb'
+require './src/services/drug_count_importer.rb'
 
-describe "Services::DrugCountsImporter" do
+describe "Services::DrugCountImporter" do
 
-  klass = Services::DrugCountsImporter
+  klass = Services::DrugCountImporter
 
 	it "should be defined" do
 		expect(klass).not_to be nil
 	end
 
-	it "should have an API for pulling drug counts by date" do
+	it "should have an API for pulling drug names" do
 		subject = klass.new
-		subject.pull(["sodium", "aspirin", "hydrochloride", "calcium"])
+		subject.pull
 		expect(subject.unsaved.size).not_to be 0
 	end
 
 	it "should be able to return data in csv format" do
 		subject = klass.new
-		subject.pull(["sodium", "aspirin", "hydrochloride", "calcium"])
+		subject.pull
 		expect(subject.to_csv).to_not be nil
 	end
 
 end
-
-
