+++ src/services/count_importer_common.rb
@@ -1,3 +1,4 @@
+
 require 'open-uri'
 require 'json'
 require 'pry'
@@ -5,30 +6,21 @@ require 'fileutils'
 
 module Services
   module CountImporterCommon
-    def get_hash url, chem
-        JSON.parse(open(url).read)["results"].each do |result|
-        @unsaved[result["time"]] = {} unless @unsaved[result["time"]]
-        @unsaved[result["time"]][chem] = result["count"]
-      end
-      true
+    def get_hash url
+			JSON.parse(open(url).read)["results"].each do |result|
+        @unsaved [result["term"]] = result["count"]
+			end	
     end
 
-    def get_csv list, name="NAME"
-      csv = []
-      if @unsaved
-        @unsaved.each do |k,values|
-          row = [k]
-          list.each do |chem|
-            row << ( values[chem] || "0" )
-          end
-          
-          csv << row.to_csv
-
-        end
-      end
-      csv.sort!
-      csv = [name + ',' + list.join(",") + "\n"] + csv
-      csv.join('')
+    def get_csv name="NAME"
+   	csv = []
+    @unsaved.each do |k, value|
+      row = [k, value]
+      csv << row.to_csv
+    end
+    csv.sort!
+    csv = [name, ", COUNT\n"] + csv
+    csv
     end
   end
 end
